Move BoringUtils to util.experimental
